+++ duckscript_sdk/src/sdk/internal/mod.rs
@@ -1,5 +1,4 @@
 mod sdkdocs;
-mod test;
 
 use duckscript::types::command::Commands;
 use duckscript::types::error::ScriptError;
@@ -8,8 +7,6 @@ static PACKAGE: &str = "internal";
 
 pub(crate) fn load(commands: &mut Commands) -> Result<(), ScriptError> {
     commands.set(sdkdocs::create(PACKAGE))?;
-
-    test::load(commands, PACKAGE)?;
 
     Ok(())
 }