+++ duckscript_sdk/src/sdk/std/collections/array_pop/mod.rs
@@ -9,7 +9,8 @@ use std::collections::HashMap;
 #[path = "./mod_test.rs"]
 mod mod_test;
 
-struct CommandImpl {
+#[derive(Clone)]
+pub(crate) struct CommandImpl {
     package: String,
 }
 
@@ -24,6 +25,10 @@ impl Command for CommandImpl {
 
     fn help(&self) -> String {
         include_str!("help.md").to_string()
+    }
+
+    fn clone_and_box(&self) -> Box<dyn Command> {
+        Box::new((*self).clone())
     }
 
     fn requires_context(&self) -> bool {