+++ duckscript_sdk/src/sdk/std/collections/write_properties/mod.rs
@@ -10,7 +10,8 @@ use std::str;
 #[path = "./mod_test.rs"]
 mod mod_test;
 
-struct CommandImpl {
+#[derive(Clone)]
+pub(crate) struct CommandImpl {
     package: String,
 }
 
@@ -25,6 +26,10 @@ impl Command for CommandImpl {
 
     fn help(&self) -> String {
         include_str!("help.md").to_string()
+    }
+
+    fn clone_and_box(&self) -> Box<dyn Command> {
+        Box::new((*self).clone())
     }
 
     fn requires_context(&self) -> bool {