+++ duckscript_sdk/src/sdk/std/env/cd/mod.rs
@@ -8,7 +8,8 @@ use std::path::Path;
 #[path = "./mod_test.rs"]
 mod mod_test;
 
-struct CommandImpl {
+#[derive(Clone)]
+pub(crate) struct CommandImpl {
     package: String,
 }
 
@@ -23,6 +24,10 @@ impl Command for CommandImpl {
 
     fn help(&self) -> String {
         include_str!("help.md").to_string()
+    }
+
+    fn clone_and_box(&self) -> Box<dyn Command> {
+        Box::new((*self).clone())
     }
 
     fn run(&self, arguments: Vec<String>) -> CommandResult {