+++ duckscript_sdk/src/sdk/std/flowcontrol/ifelse/mod.rs
@@ -289,6 +289,7 @@ fn store_call_info(call_info: &CallInfo, state: &mut HashMap<String, StateValue>
     call_info_stack.push(StateValue::SubState(call_info_state));
 }
 
+#[derive(Clone)]
 pub(crate) struct IfCommand {
     package: String,
 }
@@ -313,6 +314,10 @@ impl Command for IfCommand {
 
     fn help(&self) -> String {
         include_str!("help.md").to_string()
+    }
+
+    fn clone_and_box(&self) -> Box<dyn Command> {
+        Box::new((*self).clone())
     }
 
     fn requires_context(&self) -> bool {
@@ -385,6 +390,7 @@ impl Command for IfCommand {
     }
 }
 
+#[derive(Clone)]
 struct ElseIfCommand {
     package: String,
 }
@@ -400,6 +406,10 @@ impl Command for ElseIfCommand {
 
     fn help(&self) -> String {
         "".to_string()
+    }
+
+    fn clone_and_box(&self) -> Box<dyn Command> {
+        Box::new((*self).clone())
     }
 
     fn requires_context(&self) -> bool {
@@ -475,6 +485,7 @@ impl Command for ElseIfCommand {
     }
 }
 
+#[derive(Clone)]
 struct ElseCommand {
     package: String,
 }
@@ -490,6 +501,10 @@ impl Command for ElseCommand {
 
     fn help(&self) -> String {
         "".to_string()
+    }
+
+    fn clone_and_box(&self) -> Box<dyn Command> {
+        Box::new((*self).clone())
     }
 
     fn requires_context(&self) -> bool {
@@ -523,6 +538,7 @@ impl Command for ElseCommand {
     }
 }
 
+#[derive(Clone)]
 pub(crate) struct EndIfCommand {
     package: String,
 }
@@ -547,6 +563,10 @@ impl Command for EndIfCommand {
 
     fn help(&self) -> String {
         "".to_string()
+    }
+
+    fn clone_and_box(&self) -> Box<dyn Command> {
+        Box::new((*self).clone())
     }
 
     fn run(&self, _arguments: Vec<String>) -> CommandResult {