+++ duckscript_sdk/src/sdk/std/net/http_client/mod.rs
@@ -124,7 +124,8 @@ fn do_request(url: String, options: Options) -> CommandResult {
     }
 }
 
-struct CommandImpl {
+#[derive(Clone)]
+pub(crate) struct CommandImpl {
     package: String,
 }
 
@@ -139,6 +140,10 @@ impl Command for CommandImpl {
 
     fn help(&self) -> String {
         include_str!("help.md").to_string()
+    }
+
+    fn clone_and_box(&self) -> Box<dyn Command> {
+        Box::new((*self).clone())
     }
 
     fn run(&self, arguments: Vec<String>) -> CommandResult {