+++ duckscript_sdk/src/sdk/std/test/mod.rs
@@ -3,6 +3,8 @@ mod assert_eq;
 mod assert_error;
 mod assert_fail;
 mod assert_false;
+mod test_directory;
+mod test_file;
 
 use crate::utils::pckg;
 use duckscript::types::command::Commands;
@@ -18,6 +20,8 @@ pub(crate) fn load(commands: &mut Commands, parent: &str) -> Result<(), ScriptEr
     commands.set(assert_error::create(&package))?;
     commands.set(assert_fail::create(&package))?;
     commands.set(assert_false::create(&package))?;
+    commands.set(test_directory::create(&package))?;
+    commands.set(test_file::create(&package))?;
 
     Ok(())
 }