+++ duckscript_sdk/src/sdk/std/time/current_time/mod.rs
@@ -6,7 +6,8 @@ use std::time::{SystemTime, UNIX_EPOCH};
 #[path = "./mod_test.rs"]
 mod mod_test;
 
-struct CommandImpl {
+#[derive(Clone)]
+pub(crate) struct CommandImpl {
     package: String,
 }
 
@@ -21,6 +22,10 @@ impl Command for CommandImpl {
 
     fn help(&self) -> String {
         include_str!("help.md").to_string()
+    }
+
+    fn clone_and_box(&self) -> Box<dyn Command> {
+        Box::new((*self).clone())
     }
 
     fn run(&self, _arguments: Vec<String>) -> CommandResult {